// frontend-mobile/src/types/models.ts
export interface User {
  id: string;
  email: string;
  first_name: string;
  last_name: string;
  rpps: string;
  phone: string;
  cabinet_id: string;
  role: string;
  created_at: string;
}